[ch06-07] cleanups
