+++ lm/ch07/ASTNode.hpp
@@ -18,3 +18,4 @@ public:
 };
 
 std::string lispyTreePrint(ASTNode*);
+}
