+++ lm/ch07/ShuntingYard.cpp
@@ -1,39 +1,8 @@
 #include "ShuntingYard.hpp"
+#include "ASTNode.hpp"
 
 namespace calc
 {
-std::string lispyTreePrint(ASTNode* root)
-{
-    std::stringstream output;
-
-    if (root != nullptr)
-    {
-        if (root->data.typeId != '8' and root->data.typeId != 'S')
-        {
-            output << '(';
-            output << root->data.typeId << ' ';
-            if (root->left != nullptr)
-            {
-                output << lispyTreePrint(root->left.get()) << ' ';
-            }
-            if (root->right != nullptr)
-            {
-                output << lispyTreePrint(root->right.get());
-            }
-            output << ')';
-        }
-        else if (root->data.typeId == '8')
-        {
-            output << root->data.value;
-        }
-        else if (root->data.typeId == 'S')
-        {
-            output << root->data.id;
-        }
-    }
-    return output.str();
-}
-
 ASTBuilder::ASTBuilder(util::Log& logInit, TokenStream& tsInit)
     : log{logInit}, ts{tsInit}
 {
